% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_BASFOR_general.R
\name{run_model}
\alias{run_model}
\title{Run Basfor}
\usage{
run_model(dll = MODEL_dll, p = params, w = matrix_weather,
  calf = calendar_fert, calN = calendar_Ndep, calpT = calendar_prunT,
  caltT = calendar_thinT, n = NDAYS)
}
\description{
Run Basfor
}
\details{
This runs the Basfor Model with .,
}
\author{
a
}

